requirements.txt
requests
pandas
matplotlib
python-dotenv
seaborn
aiohttp
streamlit
sqlalchemy>=2.0
psycopg2-binary
